Some cleanup;  declare/undeclare are now functional

+++ bin/eupsLock.py
@@ -1,11 +1,16 @@
 import errno, os, stat, sys, time
 import re
 
-def lock(lockfile, who, max_wait=10):
-    """Get a lockfile, identifying yourself as who;  wait a maximum of max_wait seconds"""
+def lock(lockfile, myIdentity, max_wait=10, unlock=False, force=False):
+    """Get a lockfile, identifying yourself as myIdentity;  wait a maximum of max_wait seconds"""
+
+    myPid = os.getpid()
+    count = 0                           # count of number of times the lock is held
     while True:
         try:
             fd = os.open(lockfile, os.O_EXCL | os.O_RDWR | os.O_CREAT)
+            f = os.fdopen(fd, "w")
+            
             # we created the lockfile, so we're the owner
             break
         except OSError, e:
@@ -15,8 +20,8 @@ def lock(lockfile, who, max_wait=10):
 
         try:
             # the lock file exists, try to stat it to get its age
-            # and read it's contents to report the owner PID
-            f = open(lockfile, "r")
+            # and read its contents to report the owner and PID
+            f = os.fdopen(os.open(lockfile, os.O_EXCL | os.O_RDWR), "rw")
             s = os.stat(lockfile)
         except OSError, e:
             if e.errno != errno.EEXIST:
@@ -25,30 +30,57 @@ def lock(lockfile, who, max_wait=10):
             continue
 
         # we didn't create the lockfile and it's still there, check its age
-        pid = re.sub(r"\n$", "", f.readline())
-        who = re.sub(r"\n$", "", f.readline())
+        pid = -1
+        fileOwner = "(unknown)"
+        try:
+            pid = int(f.readline())
+            fileOwner = re.sub(r"\n$", "", f.readline())
+            count = int(f.readline())
+            f.close()
+        except:
+            pass
+
+        if pid == myPid:                # OK, we own it
+            f = open(lockfile, "w")
+            break
 
         now = int(time.time())
         if now - s[stat.ST_MTIME] > max_wait:
             raise RuntimeError, ("%s has been locked for more than %d seconds (User %s, PID %s)" %
-                                 (lockfile, max_wait, who, pid))
+                                 (lockfile, max_wait, fileOwner, pid))
 
         # it's not been locked too long, wait a while and retry
-        f.close()
-        print >> sys.stderr, "Waiting for %s (User %s, PID %s)" % (lockfile, who, pid)
+        #fd.close()
+        print >> sys.stderr, "Waiting for %s (User %s, PID %s)" % (lockfile, fileOwner, pid)
         time.sleep(2)
 
     # if we get here. we have the lockfile. Convert the os.open file
-    # descriptor into a Python file object and record our PID in it
+    # descriptor into a Python file object and record our PID, identity, and the usage count in it
+    if unlock:
+        count -= 1
 
-    f = os.fdopen(fd, "w")
-    f.write("%d\n" % os.getpid())
-    f.write("%s\n" % who)
+        if count <= 0:
+            try:
+                os.unlink(lockfile)
+            except OSError, e:
+                if e.errno != errno.ENOENT:
+                    print >> sys.stderr, "Clearing lockfile %s: %s" % (lockfile, e)
+
+            return
+    else:
+        count += 1
+
+    f.write("%d\n" % myPid)
+    f.write("%s\n" % myIdentity)
+    f.write("%d\n" % count)
     f.close()
 
-def unlock(lockfile):
-    try:
-        os.unlink(lockfile)
-    except OSError, e:
-        if e.errno != errno.ENOENT:
-            print >> sys.stderr, "Clearing lockfile %s: %s" % (lockfile, e)
+def unlock(lockfile, force=False):
+    if force:
+        try:
+            os.unlink(lockfile)
+        except OSError, e:
+            if e.errno != errno.ENOENT:
+                print >> sys.stderr, "Clearing lockfile %s: %s" % (lockfile, e)
+    else:
+        lock(lockfile, "", unlock=True)
